Made operation optional in pattern, and changes to toString(). Added logic later to handle this, with a default of ADD_NUMBER.
Spent a while wondering if this will ever lead to problems further down the line, but I don't think it will. Check this more later. The problem I was so concerned with was that the created wrapper would be incomplete (either the attribute or modifier), and so when trying to get its properties/add or remove it to an item there might be some problems.
Is this even something to worry about though? Aren't all wrappers created via this expression complete (with default values if some expressions in pattern are not provided)?